requirements.txt
anthropic[vertex]>=0.94.0
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.0.0
